requirements.txt
streamlit
google-genai>=1.0.0